requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
python-multipart>=0.0.20
python-docx>=1.1.2
Pillow>=11.0.0
